+++ tests/test_c_messaging.py
@@ -57,6 +57,23 @@ def test_delivery_rejected():
     assert str(rej_val[0][0]) == 'Failed'
     assert str(rej_val[0][1]) == 'Test failure'
 
+    error = c_uamqp.dict_value()
+    error_key = c_uamqp.string_value(b"key123")
+    error_value = c_uamqp.string_value(b"value456")
+    error[error_key] = error_value
+    error_info = c_uamqp.create_fields(error)
+
+    rej_val = c_uamqp.Messaging.delivery_rejected(b'Failed', b'Test failure', error_info)
+    assert rej_val.type == c_uamqp.AMQPType.CompositeType
+    assert rej_val.size == 1
+    assert rej_val[0].type == c_uamqp.AMQPType.CompositeType
+    assert rej_val[0].size == 3
+    assert str(rej_val[0][0]) == 'Failed'
+    assert str(rej_val[0][1]) == 'Test failure'
+    error_info = rej_val[0][2]
+    assert error_info.type == c_uamqp.AMQPType.DictValue
+    assert error_info[error_key] == error_value
+
 
 def test_delivery_released():
     rel_val = c_uamqp.Messaging.delivery_released()