+++ uamqp/errors.py
@@ -110,7 +110,6 @@ class ErrorPolicy(object):
         return ErrorAction(retry=True)
 
 
-
 class AMQPError(Exception):
     pass
 
@@ -267,7 +266,7 @@ class MessageAccepted(MessageResponse):
 
 class MessageRejected(MessageResponse):
 
-    def __init__(self, condition=None, description=None, encoding='UTF-8'):
+    def __init__(self, condition=None, description=None, encoding='UTF-8', info=None):
         if condition:
             self.error_condition = condition.encode(encoding) if isinstance(condition, six.text_type) else condition
         else:
@@ -278,6 +277,9 @@ class MessageRejected(MessageResponse):
                 else description
         else:
             self.error_description = b""
+        if info and not isinstance(info, dict):
+            raise TypeError("Disposition error info must be a dictionary.")
+        self.error_info = utils.data_factory(info, encoding=encoding) if info else None
         super(MessageRejected, self).__init__()