+++ uamqp/message.py
@@ -377,7 +377,7 @@ class Message(object):
             return True
         return False
 
-    def reject(self, condition=None, description=None):
+    def reject(self, condition=None, description=None, info=None):
         """Send a response disposition to the service to indicate that
         a received message has been rejected. If the client is running in PeekLock
         mode, the service will wait on this disposition. Otherwise it will
@@ -389,6 +389,8 @@ class Message(object):
         :type condition: bytes or str
         :param description: A description/reason to accompany the rejection.
         :type description: bytes or str
+        :param info: Information about the error condition.
+        :type info: dict
         :rtype: bool
         :raises: TypeError if the message is being sent rather than received.
         """
@@ -396,6 +398,7 @@ class Message(object):
             self._response = errors.MessageRejected(
                 condition=condition,
                 description=description,
+                info=info,
                 encoding=self._encoding)
             self._settler(self._response)
             self.state = constants.MessageState.ReceivedSettled